Masive refactor both modes up and running

+++ backend/src/main/java/org/stefan/backend/dto/PortfolioDto.java
@@ -15,6 +15,6 @@ public class PortfolioDto {
 
     private Double profit;
 
-    private int quantity;
+    private Double quantity;
 
 }
