+++ backend/src/main/java/org/stefan/backend/dto/PortfolioRequestDto.java
@@ -5,6 +5,6 @@ import java.math.BigDecimal;
 public record PortfolioRequestDto(
         BigDecimal balance,
         Double profit,
-        int quantity
+        Double quantity
 ) {
 }
